requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-restframework==0.0.1
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
